+++ src/components/search/QueryBuilder.vue
@@ -1,0 +1,7 @@
+<template>
+  <h1 />
+</template>
+
+<script setup lang="ts">
+
+</script>